version 1 of boxes&cards; v2 boxsize
boxesNcards is a data entry program used to calculate the number of
print inches for typesetting a poem and the number of hours expected to
typeset.  It is an inventory of the book of poetry "The Marriage Bed."

boxsize is now updated for a new box type - a two piece, lidded box.